+++ src/jsxer/nodes/FunctionDeclaration.cpp
@@ -4,7 +4,7 @@ namespace jsxer::nodes {
     void FunctionDeclaration::parse() {
         bodyInfo = decoders::d_line_info(reader);
         signature = decoders::d_fn_sig(reader);
-        _type = decoders::d_length(reader);
+        _type = (unsigned int) decoders::d_length(reader);
     }
 
     string FunctionDeclaration::to_string() {
